switch to ray tracing way simpler than rasterization

+++ src/rendering/framebuffer.c
@@ -3,14 +3,14 @@
 #include "allocator/allocator.h"
 #include "common/random.h"
 
-void framebuf_create(framebuf_t* buf, int32_t width, int32_t height)
+void framebuf_create(framebuf_t* buf, i32 width, i32 height)
 {
     ASSERT(buf);
     ASSERT(width > 0);
     ASSERT(height > 0);
     buf->width = width;
     buf->height = height;
-    const int32_t len = width * height;
+    const i32 len = width * height;
     buf->color = perm_malloc(len * sizeof(buf->color[0]));
     buf->depth = perm_malloc(len * sizeof(buf->depth[0]));
 }
@@ -26,24 +26,24 @@ void framebuf_destroy(framebuf_t* buf)
     buf->depth = NULL;
 }
 
-int32_t framebuf_color_bytes(framebuf_t buf)
+i32 framebuf_color_bytes(framebuf_t buf)
 {
     return buf.width * buf.height * sizeof(buf.color[0]);
 }
 
-int32_t framebuf_depth_bytes(framebuf_t buf)
+i32 framebuf_depth_bytes(framebuf_t buf)
 {
     return buf.width * buf.height * sizeof(buf.depth[0]);
 }
 
-void framebuf_clear(framebuf_t buf, uint16_t color, uint16_t depth)
+void framebuf_clear(framebuf_t buf, u32 color, float depth)
 {
-    const int32_t len = buf.width * buf.height;
-    for (int32_t i = 0; i < len; ++i)
+    const i32 len = buf.width * buf.height;
+    for (i32 i = 0; i < len; ++i)
     {
         buf.color[i] = color;
     }
-    for (int32_t i = 0; i < len; ++i)
+    for (i32 i = 0; i < len; ++i)
     {
         buf.depth[i] = depth;
     }
